Get gadget4 subfindhdf working
